move file management functions to new file for cleanliness
